accounts

+++ accounts/src/main/java/com/tss/accounts/repository/AccountRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface AccountRepository extends JpaRepository<Account, Integer> {
-    Optional<Account> findByAccountNumber(Integer accountNumber);
+    Optional<Account> findByAccountNumber(Long accountNumber);
 }
